+++ Runtime/Ingredients/Rigs/ReachPositionRig.cs
@@ -9,10 +9,12 @@ namespace GameplayIngredients.Rigs
 
         [Header("Target")]
         [SerializeField]
+        [InfoBox("Target needs to have the same parent as the current game object", InfoBoxType.Warning, "warnLocalParent")]
         protected Transform m_Target;
         [Header("Motion")]
         public float Dampen = 1.0f;
         public float MaximumVelocity = 1.0f;
+        public bool inLocalSpace = false; 
 
         [Header("On Reach Position")]
         [ReorderableList]
@@ -21,14 +23,27 @@ namespace GameplayIngredients.Rigs
 
         bool m_PositionReached = false;
 
+        bool warnLocalParent()
+        {
+            return m_Target != null && m_Target.transform.parent != transform.parent;
+        }
+
         void LateUpdate()
         {
             if(m_Target != null)
             {
                 var transform = gameObject.transform;
-                if(Vector3.Distance(transform.position , m_Target.position) < ReachSnapDistance)
+
+                Vector3 position = inLocalSpace ? transform.localPosition : transform.position;
+                Vector3 targetPosition = inLocalSpace ? m_Target.localPosition : m_Target.position;
+
+                if (Vector3.Distance(position, targetPosition) < ReachSnapDistance)
                 {
-                    transform.position = m_Target.position;
+                    if(inLocalSpace)
+                        transform.localPosition = targetPosition;
+                    else
+                        transform.position = targetPosition;
+
                     if(!m_PositionReached)
                     {
                         Callable.Call(OnReachPosition, this.gameObject);